+++ src/main/java/org/arl/fjage/remote/TunnelStatus.java
@@ -1,0 +1,9 @@
+package org.arl.fjage.remote;
+
+/**
+ * Status of a {@link Tunnel} connection.
+ */
+public enum TunnelStatus {
+  CONNECTED,
+  DISCONNECTED
+}